requirements.txt
Django>=3.2.7,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.8.6,<2.9.0

flake8>=3.9.2,<3.10.0

django-model-utils>=4.2.0,<4.3.0
django-safedelete >=1.1.0,<1.2.0
django-seed>=0.3.1,<0.4.0